Remove model inheritence (schema prep)
